MKTable Cell Bug Fixes
Added a link to the libodjc.A.dylib apple library.

Addressed issue #9
Used the runtime object association to mimic iVar for the MKControl
category that implements accessory views.

Addressed issue #10
Adjusted the calculation of accents to ensure a complete overlap of the
primary cell element. Also adjusted the layout method for accent view
trims.